Avoid implicit data type conversion caused pagefault in max_pool3d  (#4897)
When the tensor dtype is float64, the UT will pagefault because of invalid memory access. The reason may be because of the calculation overflow during the arithmetic. This fix switches every datatype to int64 to avoid the conversion.

The issue is found by the following UT:

```
export DisableScratchPages=1
export NEOReadDebugKeys=1;
cd ipex-src-folder/tests/gpu/pytorch/test/nn
pytest -sv -k "test_max_pool_nan_inf_xpu_float64" test_pooling.py 
```

The float32 test case could pass, but the float64 test fails with the following message. This indicates that the indexing calculation may have some overflow.

```
FATAL: Unexpected page fault from GPU at 0xff010087ffc00000, ctx_id: 1 (CCS) type: 0 (NotPresent), level: 4 (PML5), access: 2 (Atomic), banned: 1, aborting.
```

+++ csrc/gpu/aten/operators/DilatedMaxPool3d.cpp
@@ -25,11 +25,11 @@ struct MaxPool3dWithIndicesOutFrameImplKernelFunctor {
   void operator()(sycl::nd_item<1> item) const {
     for (auto outputIndex = item.get_global_id(0); outputIndex < OutputSize;
          outputIndex += global_range) {
-      int batch = 0;
-      int channel = 0;
-      int oDepth = 0;
-      int oRow = 0;
-      int oColumn = 0;
+      int64_t batch = 0;
+      int64_t channel = 0;
+      int64_t oDepth = 0;
+      int64_t oRow = 0;
+      int64_t oColumn = 0;
       // used only for channels-first indexing
       int64_t slice = 0;
       batch = outputIndex / out_batch_stride;
